Prefer ncursesw wide-char lib and v6; cache lib names
Raku loves Unicode; we always want the wide-char version of the library
if available.

Also, prefer ncurses 6 over 5 if both are installed.

Don't export `library` routine. It is too common a name to leak out.
Instead, declare it as `our`s, so it can be called if explicitly
qualified with the module name (`NCurses::library()`).

Refactor to remove redundant logic for looking up libraries. Now the one
`library` function can look up all four libraries (`library('menu')`
works as you'd expect).

Cache library lookup in a state variable, only do a single `library-check`
call.

Move "LibraryCheck" from "test-depends" to "depends"; it is needed for
picking the right library (and already was, that's not new in this
patch).